Chore(login.ts): checkUserExists 함수 user.ts 로 이전 함
