+++ app/services/api/index.ts
@@ -6,7 +6,6 @@ export * from "./creche-date"
 export * from "./creche"
 export * from "./creches"
 export * from "./favorite"
-export * from "./login"
 export * from "./payment-calculate"
 export * from "./payment"
 export * from "./petsitter"
